feat:클릭한 게시물 펼치기
